Move request and request_no_content to http_common
