+++ cert/aziot-cert-client-async/Cargo.toml
@@ -8,8 +8,6 @@ edition = "2018"
 http = "0.2"
 hyper = "0.14"
 percent-encoding = "2"
-serde = "1"
-serde_json = "1"
 
 aziot-cert-common-http = { path = "../aziot-cert-common-http" }
 aziot-key-common = { path = "../../key/aziot-key-common" }
