+++ key/aziot-key-client-async/Cargo.toml
@@ -8,8 +8,6 @@ edition = "2018"
 http = "0.2"
 hyper = "0.14"
 percent-encoding = "2"
-serde = "1"
-serde_json = "1"
 
 http-common = { path = "../../http-common", features = ["tokio1"] }
 aziot-key-common = { path = "../aziot-key-common" }
