+++ tpm/aziot-tpm-client-async/Cargo.toml
@@ -7,9 +7,6 @@ edition = "2018"
 [dependencies]
 http = "0.2"
 hyper = "0.14"
-percent-encoding = "2"
-serde = "1"
-serde_json = "1"
 
 aziot-tpm-common = { path = "../aziot-tpm-common" }
 aziot-tpm-common-http = { path = "../aziot-tpm-common-http" }
